+++ Editor/Drawers/MinValueDrawer/MinValueAttributeDrawer.cs
@@ -1,0 +1,103 @@
+﻿using System.Collections.Generic;
+using System.Reflection;
+using SaintsField.Editor.AutoRunner;
+using SaintsField.Editor.Core;
+using SaintsField.Editor.Utils;
+using UnityEditor;
+using UnityEngine;
+
+namespace SaintsField.Editor.Drawers.MinValueDrawer
+{
+#if ODIN_INSPECTOR
+    [Sirenix.OdinInspector.Editor.DrawerPriority(Sirenix.OdinInspector.Editor.DrawerPriorityLevel.SuperPriority)]
+#endif
+    [CustomPropertyDrawer(typeof(MinValueAttribute), true)]
+    public partial class MinValueAttributeDrawer : SaintsPropertyDrawer, IAutoRunnerFixDrawer
+    {
+        private static (string error, float valueLimit) GetLimitFloat(SerializedProperty property, MinValueAttribute minValueAttribute, MemberInfo info, object parentTarget)
+        {
+            return string.IsNullOrEmpty(minValueAttribute.ValueCallback)
+                ? ("", minValueAttribute.Value)
+                : Util.GetOf(minValueAttribute.ValueCallback, 0f, property, info, parentTarget);
+        }
+
+        public AutoRunnerFixerResult AutoRunFix(PropertyAttribute propertyAttribute, IReadOnlyList<PropertyAttribute> allAttributes,
+            SerializedProperty property, MemberInfo memberInfo, object parent)
+        {
+            if (property.isArray)
+            {
+                return null;
+            }
+
+            // ReSharper disable once SwitchStatementHandlesSomeKnownEnumValuesWithDefault
+            switch (property.propertyType)
+            {
+                case SerializedPropertyType.Float:
+                {
+                    MinValueAttribute minValueAttribute = (MinValueAttribute) propertyAttribute;
+                    (string error, float valueLimit) = GetLimitFloat(property, minValueAttribute, memberInfo, parent);
+                    if (error != "")
+                    {
+                        return new AutoRunnerFixerResult
+                        {
+                            ExecError = error,
+                            Error = "",
+                        };
+                    }
+                    if (property.floatValue < valueLimit)
+                    {
+                        return new AutoRunnerFixerResult
+                        {
+                            ExecError = "",
+                            Error = $"Expected value >= {valueLimit}, get {property.floatValue}",
+                            CanFix = true,
+                            Callback = () =>
+                            {
+                                property.floatValue = valueLimit;
+                                property.serializedObject.ApplyModifiedProperties();
+                            },
+                        };
+                    }
+
+                    return null;
+                }
+                case SerializedPropertyType.Integer:
+                {
+                    MinValueAttribute minValueAttribute = (MinValueAttribute) propertyAttribute;
+                    (string error, float valueLimit) = GetLimitFloat(property, minValueAttribute, memberInfo, parent);
+                    if (error != "")
+                    {
+                        return new AutoRunnerFixerResult
+                        {
+                            ExecError = error,
+                            Error = "",
+                        };
+                    }
+
+                    if (property.intValue < valueLimit)
+                    {
+                        return new AutoRunnerFixerResult
+                        {
+                            ExecError = "",
+                            Error = $"Expected value >= {valueLimit}, get {property.intValue}",
+                            CanFix = true,
+                            Callback = () =>
+                            {
+                                property.intValue = (int)valueLimit;
+                                property.serializedObject.ApplyModifiedProperties();
+                            }
+                        };
+                    }
+
+                    return null;
+                }
+                default:
+                    return new AutoRunnerFixerResult
+                    {
+                        ExecError = "",
+                        Error = $"Unsupported type {property.propertyType} for {property.propertyPath}",
+                    };
+            }
+        }
+    }
+}