+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## WIP ##
+
+1.  Add Auto Validator for `MinValue`, fix `MinValue` not get called on first inspecting of the field in IMGUI. Fix `MinValue` use shared error message when using on a list in IMGUI.
+2.  Same fix for `MaxValue`
+
 ## 3.32.0 ##
 
 1.  `ArraySize` now support callback argument to dynamicly limit the size.